改名為ezlaravel 架構進行簡化，並除錯中 支援Laravel8.Voyager1.4

+++ publishable/database/migrations/2017_01_15_164551_create_media_tag.php
@@ -14,9 +14,9 @@ class CreateMediaTag extends Migration
     {
         Schema::create('media_tag', function(Blueprint $table)
         {
-           $table->integer('media_id')->unsigned()->index();    
+           $table->bigInteger('media_id')->unsigned()->index();
            $table->foreign('media_id')->references('id')->on('medias')->onDelete('cascade');
-           $table->integer('tag_id')->unsigned()->index();      
+           $table->bigInteger('tag_id')->unsigned()->index();
            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
            $table->timestamps();
         });
@@ -30,7 +30,7 @@ class CreateMediaTag extends Migration
     public function down()
     {
         Schema::table('media_tag', function(Blueprint $table)
-        { 
+        {
             $table->dropForeign(['media_id']);
             $table->dropForeign(['tag_id']);
         });
