+++ publishable/database/migrations/2017_01_26_112101_create_article_tag.php
@@ -14,9 +14,9 @@ class CreateArticleTag extends Migration
     {
         Schema::create('article_tag', function(Blueprint $table)
         {
-           $table->integer('article_id')->unsigned()->index();    
+           $table->bigInteger('article_id')->unsigned()->index();
            $table->foreign('article_id')->references('id')->on('articles')->onDelete('cascade');
-           $table->integer('tag_id')->unsigned()->index();      
+           $table->bigInteger('tag_id')->unsigned()->index();
            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
            $table->timestamps();
         });
@@ -30,7 +30,7 @@ class CreateArticleTag extends Migration
     public function down()
     {
         Schema::table('article_tag', function(Blueprint $table)
-        { 
+        {
             $table->dropForeign(['article_id']);
             $table->dropForeign(['tag_id']);
         });
