+++ publishable/database/migrations/2018_04_09_124213_create_item_tag.php
@@ -15,9 +15,9 @@ class CreateItemTag extends Migration
     {
         Schema::create('item_tag', function(Blueprint $table)
         {
-            $table->integer('item_id')->unsigned()->index();
+            $table->bigInteger('item_id')->unsigned()->index();
             $table->foreign('item_id')->references('id')->on('items')->onDelete('cascade');
-            $table->integer('tag_id')->unsigned()->index();
+            $table->bigInteger('tag_id')->unsigned()->index();
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
             $table->timestamps();
         });
